requirements.txt
streamlit
pandas
nba_api
supabase
python-dotenv
transformers
torch
requests
groq